Extract sequential_slug_calculator to sequentially_slugged/calculator
